Team4

+++ app/src/app/components/homeComponent/home.component.ts
@@ -6,6 +6,7 @@ import { addincidencepageComponent } from '../addincidencepageComponent/addincid
 import { updatepageComponent } from '../updatepageComponent/updatepage.component'
 import { MatTable, MatTableDataSource } from '@angular/material/table'
 import { incidenceService } from '../../sd-services/incidenceService'
+import { NeutrinosOAuthClientService } from 'neutrinos-oauth-client'
 // import { MatDialogRef } from '@angular/material';
 
 /*
@@ -27,12 +28,13 @@ import { HeroService } from '../../services/hero/hero.service';
 export class homeComponent extends NBaseComponent implements OnInit {
     datasource
     // dataDialogRef: MatDialogRef<addincidencepageComponent>;
-    constructor(public dialog: MatDialog, public clientService: incidenceService) {
+    constructor(public dialog: MatDialog, public clientService: incidenceService, public neutrinosOAuthClientService: NeutrinosOAuthClientService) {
         super();
     }
 
     async callClientService() {
       const resultData = (await this.clientService.getdata()).local.result
+      console.log('resultDatta ', resultData)
       this.datasource = new MatTableDataSource(resultData)
     }
 
@@ -53,14 +55,16 @@ export class homeComponent extends NBaseComponent implements OnInit {
 
 
    async deleterecord(e) {
-       const reqbody = { body: {
-            "filter": {
-                 "_id": e
-         }
-       }}
-       const resultData = (await this.clientService.deletedata(reqbody))
-       console.log('result ', resultData)
-       this.ngOnInit()
+       if (confirm("Are you sure?")) {
+        const reqbody = { body: {
+                "filter": {
+                    "_id": e
+            }
+        }}
+        const resultData = (await this.clientService.deletedata(reqbody))
+        console.log('result ', resultData)
+        this.ngOnInit()
+       }
     }
 
 
@@ -73,5 +77,9 @@ export class homeComponent extends NBaseComponent implements OnInit {
         this.ngOnInit()
         console.log('close', JSON.stringify(result));
     });
+    }
+
+    logout() {
+       this.neutrinosOAuthClientService.logout('loginpage')
     }
 }
